# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProviEmplea API",
        "description": "API REST para la plataforma de empleo ProviEmplea de Providencia.\n    Gestiona talentos, empresas y procesos de selección con curriculum ciego.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local (Docker)"
        }
    ],
    "paths": {
...
        "/personas": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Listar talentos en formato CV ciego",
                "description": "Retorna perfiles activos sin datos identificatorios. **Caché 5 min.** Los empleadores NO tienen acceso a nombre, email, teléfono, edad, género ni comuna de residencia (privacidad para evitar discriminación).",
                "operationId": "listarPersonas",
                "parameters": [
                    {
                        "name": "validado",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "nivel_educacional",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "basica",
                                "media",
                                "tecnica",
                                "universitaria",
                                "postgrado"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de CV ciegos (caché 5 min, rate limit 60/min)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PersonaCVCiego"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Personas"
                ],
                "summary": "Registrar nuevo talento",
                "description": "Crea un perfil. El `codigo_talento` se genera automáticamente (PROV-YYYY-XXXX). El perfil inicia con `validado=false` y `activo=true`.",
                "operationId": "crearPersona",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Talento registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Persona"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidacion"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/personas/{id}": {
            "get": {
                "tags": [
                    "Personas"
                ],
                "summary": "Obtener perfil completo de un talento",
                "description": "Retorna **todos** los datos, incluyendo email y teléfono. Uso administrativo interno.",
                "operationId": "obtenerPersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Perfil obtenido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Persona"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "put": {
                "tags": [
                    "Personas"
                ],
                "summary": "Actualizar perfil de talento",
                "description": "Todos los campos opcionales. Recalcula `porcentaje_completitud` automáticamente.",
                "operationId": "actualizarPersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PersonaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Persona"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Personas"
                ],
                "summary": "Desactivar perfil de talento (soft delete)",
                "description": "Pone `activo=false`. No elimina el registro de la BD para preservar integridad referencial con `contactos_solicitados`.",
                "operationId": "desactivarPersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Perfil desactivado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Persona desactivada exitosamente."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/personas/{id}/validar": {
            "patch": {
                "tags": [
                    "Personas"
                ],
                "summary": "Encender o apagar visibilidad del perfil en la vitrina",
                "description": "**Toggle** del campo `validado`. Si estaba `false` lo pone `true` (encender), y si estaba `true` lo pone `false` (apagar). Solo perfiles validados aparecen en la vitrina pública de talentos. **No requiere body.**",
                "operationId": "validarPersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado de validación cambiado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Persona"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "PersonaCVCiego": {
                "title": "Persona CV Ciego",
                "description": "Perfil anónimo sin datos identificatorios.",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "codigo_talento": {
                        "type": "string",
                        "example": "PROV-2026-A1B2"
                    },
                    "resumen": {
                        "type": "string",
                        "nullable": true
                    },
                    "nivel_educacional": {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "basica",
                            "media",
                            "tecnica",
                            "universitaria",
                            "postgrado"
                        ]
                    },
                    "titulo_carrera": {
                        "type": "string",
                        "nullable": true
                    },
                    "anio_egreso": {
                        "type": "integer",
                        "nullable": true
                    },
                    "anios_experiencia": {
                        "type": "integer"
                    },
                    "areas_experiencia": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "competencias": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "rango_renta": {
                        "type": "string",
                        "nullable": true
                    },
                    "tipo_jornada": {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "completa",
                            "part-time",
                            "por-horas"
                        ]
                    },
                    "modalidad": {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "presencial",
                            "remoto",
                            "hibrido"
                        ]
                    },
                    "idiomas": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Idioma"
                        },
                        "nullable": true
                    },
                    "persona_discapacidad": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "PersonaInput": {
                "title": "PersonaInput",
                "description": "Campos para crear/actualizar talento.",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "talento@ejemplo.cl"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "+56912345678",
                        "nullable": true
                    },
                    "resumen": {
                        "type": "string",
                        "nullable": true
                    },
                    "nivel_educacional": {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "basica",
                            "media",
                            "tecnica",
                            "universitaria",
                            "postgrado"
                        ]
                    },
                    "titulo_carrera": {
                        "type": "string",
                        "nullable": true
                    },
                    "anio_egreso": {
                        "type": "integer",
                        "example": 2020,
                        "nullable": true
                    },
                    "anios_experiencia": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "areas_experiencia": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "competencias": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "rango_renta": {
                        "type": "string",
                        "example": "800k-1.2M",
                        "nullable": true
                    },
                    "tipo_jornada": {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "completa",
                            "part-time",
                            "por-horas"
                        ]
                    },
                    "modalidad": {
                        "type": "string",
                        "nullable": true,
                        "enum": [
                            "presencial",
                            "remoto",
                            "hibrido"
                        ]
                    },
                    "portafolio_url": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "persona_discapacidad": {
                        "type": "boolean",
                        "example": false,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
...
    },
...
}